+++ GJJP8B_HFT_2021221.Repository/ICheeseRepository.cs
@@ -11,7 +11,7 @@ namespace GJJP8B_HFT_2021221.Repository
     {
         void ChangeName(int id, string newName);
 
-        void ChangePrice(int id, int newPrice);
+        void ChangePrice(int id, float newPrice);
 
         void ChangeMilk(int id, int newMilkId);
     }
